refactor: moved helper class to separate file
